Ensure get_host_port tests run without Flask
